Extract process killing from custom executor

See merge request gitlab-org/gitlab-runner!1653

+++ helpers/process/killer_unix.go
@@ -5,16 +5,14 @@ package process
 import (
 	"os"
 	"syscall"
-
-	"gitlab.com/gitlab-org/gitlab-runner/common"
 )
 
 type unixKiller struct {
-	logger  common.BuildLogger
+	logger  Logger
 	process *os.Process
 }
 
-func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
+func newKiller(logger Logger, process *os.Process) killer {
 	return &unixKiller{
 		logger:  logger,
 		process: process,
@@ -22,9 +20,13 @@ func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
 }
 
 func (pk *unixKiller) Terminate() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Signal(syscall.SIGTERM)
 	if err != nil {
-		pk.logger.Errorln("Failed to send SIGTERM signal:", err)
+		pk.logger.Warn("Failed to terminate process:", err)
 
 		// try to kill right-after
 		pk.ForceKill()
@@ -32,8 +34,12 @@ func (pk *unixKiller) Terminate() {
 }
 
 func (pk *unixKiller) ForceKill() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
-		pk.logger.Errorln("Failed to force-kill:", err)
+		pk.logger.Warn("Failed to force-kill:", err)
 	}
 }
