+++ helpers/process/killer_windows.go
@@ -2,16 +2,14 @@ package process
 
 import (
 	"os"
-
-	"gitlab.com/gitlab-org/gitlab-runner/common"
 )
 
 type windowsKiller struct {
-	logger  common.BuildLogger
+	logger  Logger
 	process *os.Process
 }
 
-func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
+func newKiller(logger Logger, process *os.Process) killer {
 	return &windowsKiller{
 		logger:  logger,
 		process: process,
@@ -19,17 +17,26 @@ func NewKiller(logger common.BuildLogger, process *os.Process) Killer {
 }
 
 func (pk *windowsKiller) Terminate() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
-		pk.logger.Errorln("Failed to terminate:", err)
+		pk.logger.Warn("Failed to terminate process:", err)
 
+		// try to kill right-after
 		pk.ForceKill()
 	}
 }
 
 func (pk *windowsKiller) ForceKill() {
+	if pk.process == nil {
+		return
+	}
+
 	err := pk.process.Kill()
 	if err != nil {
-		pk.logger.Errorln("Failed to force-kill:", err)
+		pk.logger.Warn("Failed to force-kill:", err)
 	}
 }
